+++ src/main/java/com/newfit/reservation/repository/location/SidoAreaRepository.java
@@ -1,4 +1,4 @@
-package com.newfit.reservation.repository;
+package com.newfit.reservation.repository.location;
 
 import com.newfit.reservation.domain.location.SidoArea;
 import org.springframework.data.jpa.repository.JpaRepository;
